Export skipped messages
Skipped messages are messages that were in Rapid Pro but which were deliberately excluded when matching. In other words, they shouldn't have appeared in the dataset to begin with e.g. because they are duplicates. These can be archived, so we can use the data exported here as an input to the archiving script.